login feature

+++ src/main/java/com/norialertapp/repository/ProductRepo.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Repository;
  * Created by katherine_celeste on 9/30/16.
  */
 @Repository
-public interface ProductRepo extends JpaRepository<Product, Integer> {
+public interface ProductRepo extends JpaRepository<Product, Long> {
 
+    
 }
